Installers - Ubuntu: Adjustments and fixes for Ubuntu
This fixes Warty's bit where the config files were stashed in a tarball

Makes some other minor adjustments to handle the various changes and differences
across the various releases.  This includes, if Ubuntu early releases a distro,
like lucid, and it shows up here it won't actually create a config file for it
since it's missing the installers information.